requirements.txt
boto3
botocore
Pillow
watchdog
requests
flask
flask-json
